Https_Trigger/requirements.txt
azure-functions
azure-storage-blob==12.0.0
wget==3.2
requests==2.22.0
beautifulsoup4==4.8.0